마이페이지

+++ pet/src/main/java/pit/pet/Spec/SpecRepository.java
@@ -5,7 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface SpecRepository extends JpaRepository<Species, Integer> {
+public interface SpecRepository extends JpaRepository<Species, Long> {
     Optional<Species> findByName(String name);
     List<Species> findByNameContainingIgnoreCase(String keyword);
+
 }
